requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.0
pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
torch>=2.0.0
torchvision>=0.15.0
